Move and update `AllowedFile` class to `Statamic\Rules` w/ Laravel conventions.
